blob_* size_t replaced with uint64_t

+++ cvsnt/cvsnt-2.5.05.3744/ca_blobs_fs/src/fileio.h
@@ -28,7 +28,7 @@ inline bool blob_fileio_rename_file_if_nexist(const char*from, const char*to)
 extern FILE* blob_fileio_get_temp_file(std::string &fn, const char *tmp_path=nullptr, const char *mode="wb");
 
 class BlobFileIOPullData;
-BlobFileIOPullData* blobe_fileio_start_pull(const char* filepath, size_t &blob_sz);
+BlobFileIOPullData* blobe_fileio_start_pull(const char* filepath, uint64_t &blob_sz);
 const char *blobe_fileio_pull(BlobFileIOPullData* fp, uint64_t from, size_t &data_pulled);
 bool blobe_fileio_destroy(BlobFileIOPullData* fp);
